processing/requirements.txt
numpy
scipy
pyserial
matplotlib
pyqtgraph
PyQt5  
scikit-learn